fix(cli): address channel status review findings
