+++ src/lib/actions/sandbox/channel-status-config.ts
@@ -4,6 +4,7 @@
 import YAML from "yaml";
 import type { AgentDefinition } from "../../agent/defs";
 import { CLI_NAME } from "../../cli/branding";
+import { shellQuote } from "../../core/shell-quote";
 import type {
   RenderedChannelConfigParser,
   RenderedConfigSource,
@@ -22,6 +23,7 @@ import type {
 } from "../../messaging/manifest";
 import type { DiagnosticSignal } from "../../sandbox/whatsapp-diagnostics";
 import * as registry from "../../state/registry";
+import { configInputDetail, configValuesEqual } from "./channel-status-config-values";
 
 const CONFIG_STATUS_TIMEOUT_MS = 5_000;
 const channelManifestRegistry = createBuiltInChannelManifestRegistry();
@@ -39,12 +41,6 @@ type ExecRunner = (
 export type ChannelStatusConfigDeps = {
   execSandbox: ExecRunner;
 };
-
-// Inline single-quote shell quoting — config status probes only quote trusted
-// path strings derived from agent/channel manifests.
-export function quotePath(value: string): string {
-  return `'${value.replace(/'/g, "'\\''")}'`;
-}
 
 export function buildConfigStatusSignals(
   sandboxName: string,
@@ -143,11 +139,6 @@ function configInputLabel(
   return `${label} (${envKey})`;
 }
 
-function configInputDetail(value: MessagingSerializableValue | undefined): string {
-  if (value === undefined || value === null) return "not set";
-  return formatConfigValue(value);
-}
-
 type ExpectedConfigValue = {
   readonly value: MessagingSerializableValue | undefined;
   readonly detail: string;
@@ -180,16 +171,6 @@ function expectedConfigValue(
     detail: configInputDetail(undefined),
     hasValue: false,
   };
-}
-
-function formatConfigValue(value: MessagingSerializableValue): string {
-  if (typeof value === "string") return value.length === 0 ? '""' : value;
-  if (typeof value === "number" || typeof value === "boolean") return String(value);
-  if (Array.isArray(value)) {
-    if (value.length === 0) return "[]";
-    return value.map(formatConfigValue).join(", ");
-  }
-  return JSON.stringify(value);
 }
 
 interface ConfigRenderSource extends RenderedConfigVisibilityKey {
@@ -301,7 +282,7 @@ function readConfigSourceValues(
   for (const target of new Set(sources.map((source) => source.resolvedTarget))) {
     const result = deps.execSandbox(
       sandboxName,
-      `cat ${quotePath(target)}`,
+      `cat ${shellQuote(target)}`,
       CONFIG_STATUS_TIMEOUT_MS,
     );
     targetReads.set(
@@ -409,41 +390,4 @@ function compareConfigSource(
 
 function configSourceKey(source: ConfigRenderSource): string {
   return `${source.resolvedTarget}:${source.kind}:${source.key}`;
-}
-
-function configValuesEqual(
-  expected: MessagingSerializableValue | undefined,
-  actual: MessagingSerializableValue | undefined,
-): boolean {
-  if (expected === undefined || expected === null) return actual === undefined || actual === null;
-  if (actual === undefined || actual === null) return false;
-  if (Array.isArray(expected) || Array.isArray(actual)) {
-    const expectedList = listConfigValues(expected);
-    const actualList = listConfigValues(actual);
-    return (
-      expectedList.length === actualList.length &&
-      expectedList.every((value, index) => value === actualList[index])
-    );
-  }
-  const expectedBoolean = booleanConfigValue(expected);
-  const actualBoolean = booleanConfigValue(actual);
-  if (expectedBoolean !== null && actualBoolean !== null) return expectedBoolean === actualBoolean;
-  return formatConfigValue(expected) === formatConfigValue(actual);
-}
-
-function listConfigValues(value: MessagingSerializableValue): string[] {
-  const values = Array.isArray(value) ? value : String(value).split(",");
-  return values
-    .map((entry) => String(entry).trim())
-    .filter((entry) => entry.length > 0)
-    .sort();
-}
-
-function booleanConfigValue(value: MessagingSerializableValue): boolean | null {
-  if (typeof value === "boolean") return value;
-  if (typeof value !== "string") return null;
-  const normalized = value.trim().toLowerCase();
-  if (normalized === "1" || normalized === "true") return true;
-  if (normalized === "0" || normalized === "false") return false;
-  return null;
 }
